Fixed issue where original image was deleted before thumbnail creation on WP 3.5 (by removing deletion code completely); moved s3 to external file; allows disabling of admin panel by defining settings in wp-config
